+++ src/components/sdebar.js
@@ -1,0 +1,5 @@
+import createSidebar from "../helpers/categories";
+
+export default function sideBar(){
+  createSidebar();
+}
