+++ src/components/todolist.js
@@ -1,0 +1,8 @@
+import addButtonFunctionalities from "../helpers/buttonshelper"
+import displayHeaders, { displayTodos } from "../helpers/todos"
+
+export default function todoList(){
+    displayHeaders()
+    displayTodos()
+    addButtonFunctionalities()
+}
