+++ src/index.js
@@ -1,7 +1,9 @@
-import createSidebar from "./components/categories";
-import displayHeaders, { displayAllTodos, displayTodos } from './components/todos';
+import createSidebar from "./helpers/categories";
+import displayHeaders, { displayAllTodos, displayTodos } from './helpers/todos';
 import addButtonFunctionalities from './helpers/buttonshelper';
 import { welcome } from "./partials/welcome";
+import sideBar from "./components/sdebar";
+import todoList from "./components/todolist";
 
 if (!localStorage['categories']) {
   localStorage.setItem('categories', JSON.stringify([]))
@@ -11,10 +13,9 @@ if (JSON.parse(localStorage.getItem('categories')).length === 0) {
   welcome()
 }
 else{
-  displayHeaders()
-  displayTodos()
+  todoList()
 }
 
-createSidebar()
+sideBar()
 
 addButtonFunctionalities()
